Refactor about page and add featured image support
This commit refactors the about page in the src/pages/about.astro file. It updates the query to fetch the first document of type "pages" and "pageType" as "about". The page title and body are rendered using the PortableText component. Additionally, a featured image is added using the Image component from the astro:assets package. The image source is generated using the urlForImage function from the ../sanity/lib/urlForImage module. The image is displayed with different widths and sizes based on the viewport width. The alt attribute is set to the page title and the image is given the class "page__cover".

---

Update index page and fix post link

This commit updates the index page in the src/pages/index.astro file. It modifies the query to fetch the first three documents of type "post". The page title, body, and featured image are rendered using the PortableText component, Image component, and urlForImage function respectively. The post links are fixed to include the correct slug.