Refactor font styling and update color scheme
Update font styling in global.scss to use the primary color for font color. Also, update the dark mode font color to #FF2E16.

Update font size in Footer component

Increase the font size of the paragraph in the Footer component to var(--fs--2) and set the font family to var(--font-family-mono).